backend/requirements.txt
fastapi
uvicorn

#for template  #new
jinja2 

#for static files
aiofiles

#for database
sqlalchemy
psycopg2-binary

#for loading environment variables
python-dotenv

#for email validation
pydantic[email]

#for hashing passwords
passlib[bcrypt]

#for testing
pytest
requests

#seedir to print tree structure of folders
seedir

#for jwt tokens
python-jose